+++ src/utils/libs/getSaying.ts
